HOLY $#!7 MY DRIVE FAILED THIS IS WHAT I RECOVERED
THAT COULD HAVE GONE MUCH, MUCH WORSE.

MY ART FILES ARE SAFE. ONLY IMPORTANT THING WE LOST WAS SAPIENT QUEEN'S ANIMATIONS.

Im not sleeping easy tonight after experiencing this, making a deal with the devil (whoever owns that google drive, IOU), waiting 2 hours, staying up past midnight, just to recover 2+ years of work from multiple physical locations that was stored on a USB that decided it would pull a funny and convert itself to RAW but instead of typing /j it typed /srs so here we are, with me ugly crying in the corner knowing that it could've gone down like my first corrupted USB flash drive.

This is a cry for help. This isn't the first time my PC made a USB flash drive go FUBAR