Step 2: most blocks (NOT SAFE FOR OLD WORLDS)
Doing creeped grass for later. I'd say this is a job well-done, and these have been some of the strongest commits I've ever sent on GitHub.

I didn't need to bring out any docs or references for this. This is brilliant!

It does break some old worlds a bit by removing three blocks, but the functionality of hive creep remains untouched.